refactor(accounts): cardless mobile tab, sort menu in toolbar, drop debug panel
- Mobile Accounts tab no longer wraps content in a Card (account cards
  inside the carousel already provide card chrome); it renders a slim
  header row (rithmic refresh + charts/table toggle), the banner, and
  the carousel/table directly
- Extract the sorting popover into a reusable AccountsSortMenu with
  header and toolbar variants; desktop header uses it, and the mobile
  toolbar gains a sort button (with active-rule count badge)
- Remove the filter button from the mobile accounts toolbar
  (order: search, sort, create group, edit groups)
- Remove the Rithmic balances debug panel and its component file
